% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosmooth-package.R
\docType{package}
\name{ecosmooth-package}
\alias{ecosmooth-package}
\alias{ecosmooth}
\title{ecosmooth: Kernel-based smoothing of multivariate ecological data}
\description{
Performs smoothing of multivariate ecological data on the basis of Gaussian kernels
}
\author{
\strong{Maintainer}: Miquel De Cáceres
\email{miquelcaceres@gmail.com}
\href{https://orcid.org/0000-0001-7132-2080}{ORCID}
}
\keyword{internal}
