% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelsmoothing.R
\name{kernelsmoothing}
\alias{kernelsmoothing}
\alias{kernelmatrix}
\title{Kernel-based smoothing of multivariate data}
\usage{
kernelsmoothing(x, dist_kernel, kernel_scale)

kernelmatrix(dist_kernel, kernel_scale, normalize = TRUE)
}
\arguments{
\item{x}{A rectangular matrix (with observations in rows and variables in columns) or an object of class \code{\link{dist}} (with resemblance values between observations).}

\item{dist_kernel}{An object of class \code{\link{dist}} containing distances in the kernel space.}

\item{kernel_scale}{Bandwidth of the Gaussian kernel.}

\item{normalize}{A boolean flag to normalize kernel values to sum one for each row}
}
\value{
\itemize{
\item{Function \code{kernelsmoothing} returns either a rectangular matrix (with observations in rows and variables in columns) or an object of class \code{\link{dist}} (with resemblance values between observations),
depending on the input \code{x}.}
\item{Function \code{kernelmatrix} returns a square numeric matrix with kernel values.}
}
}
\description{
Kernel-based smoothing of multivariate data
}
\examples{
# Create original data
x <- matrix(rnorm(100), nrow = 5)

# Original distance matrix
d <- dist(x)

# Kernel matrix based on d space (bandwidth = 2)
k <- kernelmatrix(d, 2)
k

# Self-smoothing of distance data (bandwidth = 2)
kernelsmoothing(d, d, 2)

# Check equivalence with self-smoothing of rectangular data
dist(kernelsmoothing(x,d,2))
}
